LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_capture_suppressing_mouse-expected.txt



PASS Validate pointer events track pointer movement without pointer capture.
FAIL Test pointer capture. assert_equals: Check after hover on capture target: Unexpected events pointerout@target0, pointerleave@target0 expected 0 but got 2
